+++ src/main/java/proteomics/Types/ChainEntry.java
@@ -5,13 +5,13 @@ import java.util.*;
 public class ChainEntry {
 
     public final String seq;
-    public final float chain_mass;
+    public final double chain_mass;
     public final Set<Short> link_site_set;
     public final boolean n_term;
     public final boolean c_term;
     public final int binaryModType;
 
-    public ChainEntry(String seq, float chain_mass, Set<Short> link_site_set, boolean n_term, boolean c_term, int binaryModType) {
+    public ChainEntry(String seq, double chain_mass, Set<Short> link_site_set, boolean n_term, boolean c_term, int binaryModType) {
         this.seq = seq;
         this.chain_mass = chain_mass;
         this.link_site_set = link_site_set;
